Final commision, submision worked

+++ src/MP3Track.cpp
@@ -34,7 +34,7 @@ void MP3Track::analyze_beatgrid() {
     // NOTE: Use exactly 2 spaces before each arrow (→) character
         double esitmated_beats= (duration_seconds/60.0)*bpm;
         double precision_factor= bitrate/320.0;
-        std::cout << "  → Estimated beats: " << esitmated_beats 
+        std::cout << "  → Estimated beats: " << (int) esitmated_beats 
               << "  → Compression precision factor: " << precision_factor << std::endl;
 }
